+++ packages/crosswise-uikit/src/components/Card/Card.tsx
@@ -5,7 +5,12 @@ import { CardProps } from "./types";
 const Card: React.FC<CardProps> = ({ ribbon, children, background, ...props }) => {
   return (
     <StyledCard {...props}>
-      <StyledCardInner background={background} hasCustomBorder={!!props.borderBackground}>
+      <StyledCardInner
+        background={background}
+        hasCustomBorder={!!props.borderBackground}
+        hasGradientBack={!!props.isGradient}
+        isBeta={!!props.isBeta}
+      >
         {ribbon}
         {children}
       </StyledCardInner>