+++ packages/crosswise-uikit/src/components/Card/StyledCard.tsx
@@ -47,6 +47,7 @@ export const StyledCard = styled.div<StyledCardProps>`
   color: ${({ theme, isDisabled }) => theme.colors[isDisabled ? "textDisabled" : "text"]};
   overflow: hidden;
   position: relative;
+  box-shadow: ${({ theme }) => theme.card.boxShadow};
 
   ${({ isActive }) =>
     isActive &&
@@ -60,11 +61,22 @@ export const StyledCard = styled.div<StyledCardProps>`
   ${space}
 `;
 
-export const StyledCardInner = styled(Box)<{ background?: string; hasCustomBorder: boolean }>`
+export const StyledCardInner = styled(Box)<{
+  background?: string;
+  hasCustomBorder: boolean;
+  hasGradientBack?: boolean;
+  isBeta?: boolean;
+}>`
   width: 100%;
   height: 100%;
   overflow: ${({ hasCustomBorder }) => (hasCustomBorder ? "initial" : "inherit")};
-  background: ${({ theme, background }) => background ?? theme.card.background};
+  background: ${({ hasGradientBack, isBeta, theme, background }) => {
+    if (hasGradientBack) {
+      return !isBeta ? theme.card.gradientAlt : theme.card.gradientBeta;
+    }
+    return background ?? theme.card.background;
+  }};
+  // background: ${({ theme, background }) => background ?? theme.card.background};
   border-radius: ${({ theme }) => theme.radii.card};
 `;