+++ packages/crosswise-uikit/src/components/Card/theme.ts
@@ -4,8 +4,10 @@ import { CardTheme } from "./types";
 
 export const light: CardTheme = {
   background: lightColors.backgroundAlt,
+  gradientAlt: lightColors.gradients.gradprimary,
+  gradientBeta: lightColors.gradients.gradsecondary,
   boxShadow:
-    "8px 8px 24px 0 rgba(9, 13, 20, 0.4), -4px -4px 8px 0 rgba(224, 224, 255, 0.04), 0 1px 1px 0 rgba(9, 13, 20, 0.4)",
+    "8px 8px 24px 0 rgba(9, 13, 20, 0.06), -4px -4px 8px 0 rgba(255, 255, 255, 0.4), 0 1px 1px 0 rgba(9, 13, 20, 0.06)",
   // boxShadow: "solid 1px rgba(245, 247, 250, 0.06)",
   boxShadowActive: shadows.active,
   boxShadowSuccess: shadows.success,
@@ -20,6 +22,8 @@ export const light: CardTheme = {
 
 export const dark: CardTheme = {
   background: darkColors.backgroundAlt,
+  gradientAlt: darkColors.gradients.gradprimary,
+  gradientBeta: darkColors.gradients.gradsecondary,
   boxShadow:
     "8px 8px 24px 0 rgba(9, 13, 20, 0.4), -4px -4px 8px 0 rgba(224, 224, 255, 0.04), 0 1px 1px 0 rgba(9, 13, 20, 0.4)",
   // boxShadow: "solid 1px rgba(245, 247, 250, 0.06)",